Metodos read, create e delete funcionando

+++ src/main/java/br/com/fecaf/repository/VeiculoRepository.java
@@ -3,6 +3,8 @@ package br.com.fecaf.repository;
 import br.com.fecaf.model.Veiculo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface VeiculoRepository extends JpaRepository<Veiculo, Integer> {
+import java.util.List;
+
+public interface VeiculoRepository extends JpaRepository<Veiculo, Long> {
 
 }
